ai_ml_services/nlu_service/requirements.txt
grpcio
grpcio-tools
protobuf
google-cloud-dialogflow-cx>=1.0.0
python-dotenv
